+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Add `pauseSession()` and `resumeSession()` methods to `Client` [#666](https://github.com/bugsnag/bugsnag-js/pull/666)
 - Remove `client.request` property [#672](https://github.com/bugsnag/bugsnag-js/pull/672)
 - Remove `client.device` property [#673](https://github.com/bugsnag/bugsnag-js/pull/673)
+- Stop applying default error class/message when none is supplied [#676](https://github.com/bugsnag/bugsnag-js/pull/676)
 
 ## 6.5.0 (2019-12-16)