+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Remove `client.device` property [#673](https://github.com/bugsnag/bugsnag-js/pull/673)
 - Stop applying default error class/message when none is supplied [#676](https://github.com/bugsnag/bugsnag-js/pull/676)
 - Remove Bugsnag* prefix from internal class names [#679](https://github.com/bugsnag/bugsnag-js/pull/679)
+- Remove `client.app` property [#677](https://github.com/bugsnag/bugsnag-js/pull/677)
 
 ## 6.5.0 (2019-12-16)