+++ CHANGELOG.md
@@ -27,6 +27,8 @@
 - Remove `client.app` property [#677](https://github.com/bugsnag/bugsnag-js/pull/677)
 - Rename and make private the `Session` method `trackError()` -> `_track()` [#675](https://github.com/bugsnag/bugsnag-js/pull/675)
 - Update `Event` to support multiple errors [#680](https://github.com/bugsnag/bugsnag-js/pull/680)
+- Move breadcrumbs to a private property on `client._breadcrumbs` [#681](https://github.com/bugsnag/bugsnag-js/pull/681)
+- Move context to a private property on `Client`, and get/set via `getContext()/setContext()` [#681](https://github.com/bugsnag/bugsnag-js/pull/681)
 
 ## 6.5.0 (2019-12-16)