+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - Move breadcrumbs to a private property on `client._breadcrumbs` [#681](https://github.com/bugsnag/bugsnag-js/pull/681)
 - Move context to a private property on `Client`, and get/set via `getContext()/setContext()` [#681](https://github.com/bugsnag/bugsnag-js/pull/681)
 - Update `@bugsnag/safe-json-stringify` to replace redacted values with `[REDACTED]` [#683](https://github.com/bugsnag/bugsnag-js/pull/683)
+- Refactor type definitions [#682](https://github.com/bugsnag/bugsnag-js/pull/682)
 
 ## 6.5.0 (2019-12-16)