+++ CHANGELOG.md
@@ -32,6 +32,9 @@
 - Update `@bugsnag/safe-json-stringify` to replace redacted values with `[REDACTED]` [#683](https://github.com/bugsnag/bugsnag-js/pull/683)
 - Refactor type definitions [#682](https://github.com/bugsnag/bugsnag-js/pull/682)
 
+### Fixed
+- (expo): Pin `@react-native-community/netinfo` dependency to exact version bundled by Expo [#691](https://github.com/bugsnag/bugsnag-js/pull/691)
+
 ## 6.5.0 (2019-12-16)
 
 ### Added